requirements.txt
# Core dependencies
psycopg2-binary>=2.9.9
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pandas>=2.1.0
spacy>=3.7.0
rapidfuzz>=3.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
ruff>=0.1.0

